# specs/001-develop-a-platform/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Roblox Trading Platform API
  description: API for secure Roblox item trading with escrow
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /trades/{tradeId}/join:
    post:
      summary: Join trade as participant
      tags: [Trades]
      security:
        - bearerAuth: []
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Joined trade
        '400':
          description: Cannot join trade
        '401':
          description: Unauthorized

  # Chat
  /trades/{tradeId}/messages:
    get:
      summary: Get trade chat messages
      tags: [Chat]
      security:
        - bearerAuth: []
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          description: Message timestamp for pagination
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Chat messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'

    post:
      summary: Send chat message
      tags: [Chat]
      security:
        - bearerAuth: []
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'
        '400':
          description: Invalid message
        '403':
          description: Not a trade participant

  # Payments
  /trades/{tradeId}/payment-intent:
    post:
      summary: Create payment intent
      tags: [Payments]
      security:
        - bearerAuth: []
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Payment intent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  clientSecret:
                    type: string
                  paymentId:
                    type: string
        '400':
          description: Cannot create payment for this trade
        '403':
          description: Not authorized to pay for this trade
...
  /middleman/auth:
    post:
      summary: Middleman authentication
      tags: [Middleman]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, password]
              properties:
                userId:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  middlemanId:
                    type: string
        '401':
          description: Invalid credentials
...
  /middleman/trades/{tradeId}/complete:
    post:
      summary: Mark trade as completed
      tags: [Middleman]
      security:
        - bearerAuth: []
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved]
              properties:
                approved:
                  type: boolean
                notes:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Trade completed
        '403':
          description: Not the assigned middleman

  # File uploads
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar:
          type: string
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        tradeId:
          type: string
        userId:
          type: string
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'